Enable running superlesson directly
In moving main.py to __main__.py, we can now execute the module
directly.
