+++ pyproject.toml
@@ -8,9 +8,13 @@ authors = [
     "Isaque Alves",
     "Ricardo Saraiva",
 ]
+packages = [{ include = "superlesson" }]
 readme = "README.md"
 license = "MIT"
 repository = "https://github.com/SuperLesson/SuperLesson"
+
+[tool.poetry.scripts]
+superlesson = "superlesson:main"
 
 [tool.poetry.dependencies]
 python = "^3.9"
